requirements.txt
tqdm
pyyaml
loguru
DBUtils

# GUI
pyside6
PySide6-Fluent-Widgets[full]

# gpx2csv.py
gpxpy
geopandas
shapely

# svg_gen.py
svgwrite
svgpathtools @ https://github.com/bcwhite-code/svgpathtools/archive/refs/heads/master.zip
fontTools

# create_pic.py
# note that you need to install cairo to use cairosvg
cairosvg
pillow
moviepy
numpy
imageio

# fontforge_export_svg.py
# need to run in fontforge python environment

# server
fastapi
uvicorn[standard]
python-multipart
beanie
bunnet
motor
celery
redis
# windows 上要安装，https://blog.csdn.net/showgea/article/details/109342664
eventlet
flower